fix review form to close on submission fix featured responsiveness fix navbar
