+++ client/src/components/Nav/Nav.css
@@ -63,7 +63,7 @@ header {
 }
 
 
-@media only screen and (min-width: 1024px) {
+@media only screen and (min-width: 1025px) {
   .fa-bars {
     display: none;
   }
