+++ models/wine.js
@@ -24,9 +24,9 @@ const Wine = new Schema(
 Wine.virtual("rating").get(function () {
   //taking average of reviews
   return (
-    Math.ceil(this.reviews.reduce((total, review) => total + review.rating, 0) /
+    (this.reviews.reduce((total, review) => total + review.rating, 0) /
     this.reviews.length
-  ));
+  ).toFixed(2));
 });
 
 export default mongoose.model("wines", Wine);
